[refactor] move zch_to_dynamicemb_convert into tzrec/tools/dynamicemb/
The converter lives alongside the existing
tzrec/tools/dynamicemb/create_dynamicemb_init_ckpt.py instead of at the
top of tzrec/tools/. Update the test's `as conv` import, the
subprocess command in ConvertE2ETests, and the user-doc invocation in
docs/source/feature/dynamicemb.md.

All 33 tests still pass (~40s).
