+++ docs/source/feature/dynamicemb.md
@@ -97,10 +97,10 @@ feature_configs {
 
 ## 从ZCH训练好的模型迁移
 
-如果已有用 `zch{}` 训练好的模型，希望切换到 `dynamicemb{}` 继续训练，可以使用 `tzrec.tools.zch_to_dynamicemb_convert` 将ZCH checkpoint 转换为 dynamicemb 格式的 checkpoint，下游训练任务直接通过 `--fine_tune_checkpoint` 热启动。
+如果已有用 `zch{}` 训练好的模型，希望切换到 `dynamicemb{}` 继续训练，可以使用 `tzrec.tools.dynamicemb.zch_to_dynamicemb_convert` 将ZCH checkpoint 转换为 dynamicemb 格式的 checkpoint，下游训练任务直接通过 `--fine_tune_checkpoint` 热启动。
 
 ```bash
-python -m tzrec.tools.zch_to_dynamicemb_convert \
+python -m tzrec.tools.dynamicemb.zch_to_dynamicemb_convert \
 --source_checkpoint_path {ZCH_CKPT_PATH}/model.ckpt-N \
 --source_pipeline_config_path {PATH_TO_CONFIG_WITH_ZCH} \
 --target_pipeline_config_path {PATH_TO_CONFIG_WITH_DYNAMICEMB} \
