+++ tzrec/tools/dynamicemb/zch_to_dynamicemb_convert.py
